# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AUG22pt-ep1-dissolution2",
    "description": "<b>Swagger Doc EP Noroff Backend API</b>."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "scheme": "bearer",
      "description": "Bearer token"
    }
  },
  "paths": {
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Get all Categorys in the database",
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { result: \"category found\", category }"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Admin can Add a Category to the database [requires: token]",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CategoryPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"New category has been added to the database\", input: categoryName }"
          },
          "403": {
            "description": "statusCode: 401, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode: 500, result: {error: \"categoryName allready exist\", input: categoryName }"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Update/Change a Category to the database",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CategoryPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"categoryName has been updated\", input: newCategoryName }"
          },
          "403": {
            "description": "statusCode: 401, result: { error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode: 500, result: { error: \"categoryName dont exist\", input: categoryName }"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Delete a Category from the database",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CategoryDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"categoryName has been deleted\", input: categoryName }"
          },
          "403": {
            "description": "statusCode: 403, result: { error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: { error: \"category is associated with products, restricted for delete\", input: categoryName }"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CategoryPost": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string",
          "example": "categoryName"
        }
      },
      "required": [
        "categoryName"
      ]
    },
    "CategoryPut": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string",
          "example": "categoryName"
        },
        "newCategoryName": {
          "type": "string",
          "example": "newCategoryName"
        }
      },
      "required": [
        "categoryName",
        "newCategoryName"
      ]
    },
    "CategoryDelete": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string",
          "example": "newCategoryName"
        }
      },
      "required": [
        "categoryName"
      ]
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}